Main Functions

Nearest Correlation Matrix

MvSim.nearestPSDcor — Method

nearestPSDcor(R;
    τ::Real=1e-5,
    iter_outer=200,
    iter_inner=20,
    N=200,
    δ::Real=1e-6,
    ϵ::Real=1e-2,
    σ::Real=1e-4)

Compute the nearest positive semidefinite correlation matrix given a symmetric correlation matrix R. This algorithm is based off of work by Qi and Sun 2006. Matlab, C, R, and Python code can be found on Sun's page. The algorithm has also been implemented in Fortran in the NAG library.

Arguments

τ::Real: a [small] nonnegative number used to enforce a minimum eigenvalue.

δ::Real: the error tolerance for the stopping condition.

Examples

import LinearAlgebra: eigvals
# Define a negative definite correlation matrix
ρ = [1.00 0.82 0.56 0.44
     0.82 1.00 0.28 0.85
     0.56 0.28 1.00 0.22
     0.44 0.85 0.22 1.00]
eigvals(ρ)

r = nearestPSDcor(ρ)
eigvals(r)

Utilities

MvSim.gradient — Method

gradient(y, λ₀, P, b₀)

MvSim.PCA — Method

PCA(X, λ, P)

MvSim.pre_cg — Method

pre_cg(b, c, Ω₀, P, ϵ, N)

Pre- Conjugate Gradient method.

MvSim.precond_matrix — Method

precond_matrix(Ω₀, P)

MvSim.set_omega — Method

set_omega(λ)

MvSim.jacobian — Method

jacobian(x, Ω₀, P; PERTURBATION=1e-9)

Pearson Matching

Continuous

MvSim.ρz — Function

ρz(ρx, dA::ContinuousUnivariateDistribution, dB::ContinuousUnivariateDistribution, μA, μB, σA, σB, n::Int=3)

Estimate the input correlation coefficient ρz given the marginal CDFs of two continuous univariate distributions and the desired correlation coefficient ρx.

MvSim.ρz — Function

ρz(ρx, dA::ContinuousUnivariateDistribution, dB::ContinuousUnivariateDistribution, n::Int=3)

Estimate the input correlation coefficient ρz given the marginal CDFs of two continuous univariate distributions and the desired correlation coefficient ρx. Calculate the mean and variance for each distribution.

Discrete

MvSim.ρz — Function

ρz(ρx, dA::DiscreteUnivariateDistribution, dB::DiscreteUnivariateDistribution, σA, σB, minA, minB, maxA, maxB, n::Int=3)

Estimate the input correlation coefficient ρz given the marginal CDFs of two discrete univariate distributions and the desired correlation coefficient ρx.

MvSim.ρz — Function

ρz(ρx, dA::DiscreteUnivariateDistribution, dB::DiscreteUnivariateDistribution, n::Int=3)

Estimate the input correlation coefficient ρz given the marginal CDFs of two discrete univariate distributions and the desired correlation coefficient ρx. Calculate the variance, minimum, and maximum for each distribution.

Mixed

MvSim.ρz — Function

ρz(ρx, dA::DiscreteUnivariateDistribution, dB::ContinuousUnivariateDistribution, σA, σB, minA, maxA, n::Int=3)

Estimate the input correlation coefficient ρz given the marginal CDFs of two mixed support univariate distributions and the desired correlation coefficient ρx.

MvSim.ρz — Function

ρz(ρx, dA::DiscreteUnivariateDistribution, dB::ContinuousUnivariateDistribution, n::Int=3)

Estimate the input correlation coefficient ρz given the marginal CDFs of two mixed support univariate distributions and the desired correlation coefficient ρx. Calculate the support as needed.

MvSim.ρz — Function

ρz(ρx, dA::ContinuousUnivariateDistribution, dB::DiscreteUnivariateDistribution, σA, σB, minB, maxB, n::Int=3)

Estimate the input correlation coefficient ρz given the marginal CDFs of two mixed support univariate distributions and the desired correlation coefficient ρx.

MvSim.ρz — Function

ρz(ρx, dA::ContinuousUnivariateDistribution, dB::DiscreteUnivariateDistribution, n::Int=3)

Estimate the input correlation coefficient ρz given the marginal CDFs of two mixed support univariate distributions and the desired correlation coefficient ρx. Calculate the support as needed.

Utilities

MvSim.get_coefs — Method

get_coefs(margin::UnivariateDistribution, n::Int)

Get the $n^{th}$ degree Hermite Polynomial expansion coefficients for $F^{-1}[Φ(⋅)]$ where $F^{-1}$ is the inverse CDF of a probability distribution and Φ(⋅) is the CDF of a standard normal distribution.

Notes

The paper describes using Guass-Hermite quadrature using the Probabilists' version of the Hermite polynomials, while the package FastGaussQuadrature.jl uses the Physicists' version. Because of this, we need to do a rescaling of the input and the output:

\[\frac{1}{k!}\sum_{s=1}^{m}w_s H_k (t_s) F_{i}^{-1}\left[\Phi(t_s)\right] ⟹ \frac{1}{\sqrt{\pi} \cdot k!}\sum_{s=1}^{m}w_s H_k (t_s\sqrt{2}) F_{i}^{-1}\left[\Phi(t_s)\right]\]

MvSim.Gn0d — Method

Gn0d(::Int, A, B, α, β, σAσB_inv)

Calculate the $n^{th}$ derivative of G at 0 where $ρ_x = G(ρ_z)$

We are essentially calculating a double integral over a rectangular region

\[\int_{α_{r-1}}^{α_r} \int_{β_{s-1}}^{β_s} Φ(z_i, z_j, ρ_z) dz_i dz_j\]

(α[r], β[s+1]) +----------+ (α[r+1], β[s+1])
               |          |
               |          |
               |          |
  (α[r], β[s]) +----------+ (α[r+1], β[s])

MvSim.Gn0m — Method

Gn0m(::Int, A, α, dB, σAσB_inv)

Calculate the $n^{th}$ derivative of G at 0 where $ρ_x = G(ρ_z)$

MvSim.ρz_bounds — Method

ρz_bounds(::UnivariateDistribution, ::UnivariateDistribution; ::Int=3)

Compute the lower and upper bounds of possible correlations for a pair of univariate distributions. The value n determines the accuracy of the approximation of the two distributions.

MvSim.solvePoly_pmOne — Method

solvePoly_pmOne(coef)

Solve a polynomial equation on the interval [-1, 1].