Utilities

MvSim.cor2cor — Method

cor2cor(ρ::T, from::Symbol, to::Symbol) where {T <: Real}

Convert from one type of correlation matrix to another. The possible correlation types are Pearson (:P), Spearman (:S), or Kendall (:K). If an invalid pair is given, throw an error.

MvSim.cor2cor — Method

cor2cor(A::AbstractMatrix{T}, from::Symbol, to::Symbol) where {T <: Real}

Convert from one type of correlation matrix to another. The possible correlation types are Pearson (:P), Spearman (:S), or Kendall (:K). If an invalid pair is given, throw an error.

MvSim.cov2cor — Method

cov2cor(Σ::AbstractArray)

Convert a covariance matrix to a correlation matrix. Ensure that the resulting matrix is symmetric and has diagonals equal to 1.0.

MvSim.hermite — Function

hermite(x, n::Int, probabilists::Bool=true)

Compute the Hermite polynomials of degree n. Compute the Probabilists' version by default.

The two definitions of the Hermite polynomials are each a rescaling of the other. Let $Heₙ(x)$ denote the Probabilists' version, and $Hₙ(x)$ the Physicists'. Then

\[H_{n}(x) = 2^{\frac{n}{2}} He_{n}\left(\sqrt{2} x\right)\]

\[He_{n}(x) = 2^{-\frac{n}{2}} H_{n}\left(\frac{x}{\sqrt{2}}\right)\]

MvSim.rcor — Function

rcor(d::Integer, α::Real=1.0)

Generate a random positive definite correlation matrix of size $d×d$. The parameter α is used to determine the autocorrelation in the correlation coefficients.

Reference

Joe H (2006). Generating random correlation matrices based on partial correlations. J. Mult. Anal. Vol. 97, 2177–2189.

MvSim.setdiag — Method

setdiag(A::AbstractMatrix{T}, x::S) where {T<:Real, S<:Real}

Set the diagonal elements of an AbstractMatrix to a value. Return the new matrix.

MvSim.z2x — Method

z2x(d::UnivariateDistribution, x::AbstractArray)

Convert samples from a standard normal distribution to a given marginal distribution.