Main Functions

Random Multivariate Vector

MvSim.MvDistribution — Type

MvDistribution(R, margins, C)

Simple data structure for storing a multivariate mixed distribution.

MvSim.rvec — Function

rvec(n, margins, ρ)

Generate samples for a list of marginal distributions and a correaltion structure.

Base.rand — Method

Base.rand(D::MvDistribution, n::Int)

More general wrapper for rvec.

Correlations

Statistics.cor — Function

cor(x, ::Type{<:Correlation})

Compute the correlation matrix. The possible correlation types are Pearson, Spearman, or Kendall.

MvSim.cor_convert — Function

cor_convert(ρ::Real, from::Correlation, to::Correlation)

Convert from one type of correlation matrix to another. The possible correlation types are Pearson, Spearman, or Kendall.

MvSim.cor_nearPD — Method

cor_nearPD(R::Matrix{Float64};
    τ::Float64=1e-5,
    iter_outer::Int=200,
    iter_inner::Int=20,
    N::Int=200,
    err_tol::Float64=1e-6,
    precg_err_tol::Float64=1e-2,
    newton_err_tol::Float64=1e-4)

Compute the nearest positive definite correlation matrix given a symmetric correlation matrix R. This algorithm is based off of work by Qi and Sun 2006. Matlab, C, R, and Python code can be found on Sun's page. The algorithm has also been implemented in Fortran in the NAG library.

Arguments

τ::Float64: a [small] nonnegative number used to enforce a minimum eigenvalue.

err_tol::Float64: the error tolerance for the stopping condition.

Examples

import LinearAlgebra: eigvals
# Define a negative definite correlation matrix
ρ = [1.00 0.82 0.56 0.44
     0.82 1.00 0.28 0.85
     0.56 0.28 1.00 0.22
     0.44 0.85 0.22 1.00]
eigvals(ρ)

r = cor_nearPD(ρ)
eigvals(r)

MvSim.cor_nearPSD — Method

cor_nearPSD(A::Matrix{T}; n_iter::Int=100) where {T<:Real}

Examples

import LinearAlgebra: eigvals
# Define a negative definite correlation matrix
ρ = [
    1.00 0.82 0.56 0.44
    0.82 1.00 0.28 0.85
    0.56 0.28 1.00 0.22
    0.44 0.85 0.22 1.00
]
eigvals(ρ)

r = cor_nearPSD(ρ, n_iter=100)
eigvals(r)

MvSim.cor_randPD — Function

cor_randPD(T::Type{<:AbstractFloat}, d::Int, α::Real=1.0)

Generate a random positive definite correlation matrix of size $d×d$. The parameter α is used to determine the autocorrelation in the correlation coefficients.

Reference

Joe H (2006). Generating random correlation matrices based on partial correlations. J. Mult. Anal. Vol. 97, 2177–2189.

MvSim.cor_randPSD — Function

cor_randPSD(T::Type{<:AbstractFloat}, d::Int, k::Int=d)

Compute a random positive semidefinite correlation matrix

Reference

https://stats.stackexchange.com/a/125020

Pearson Matching

MvSim.ρz — Method

ρz(ρx::Real, dA::UD, dB::UD; n::Int=7)

Compute the pearson correlation coefficient that is necessary to achieve the target correlation given a pair of marginal distributions.

MvSim.ρz_bounds — Function

ρz_bounds

Compute the lower and upper bounds of possible correlations for a pair of univariate distributions. The value n determines the accuracy of the approximation of the two distributions.