Main Functions

MvSim.nearestPSDcor — Method

nearestPSDcor(R)

Compute the nearest positive semidefinite correlation matrix given a symmetric correlation matrix R. This algorithm is based off of work by Qi and Sun 2006. Matlab, C, R, and Python code can be found on Sun's page. The algorithm has also been implemented in Fortran in the NAG library.

Arguments

τ::Real: a [small] nonnegative number used to enforce a minimum eigenvalue.

δ::Real: the error tolerance for the stopping condition.

Examples

import LinearAlgebra: eigvals
# Define a negative definite correlation matrix
ρ = [1.00 0.82 0.56 0.44
     0.82 1.00 0.28 0.85
     0.56 0.28 1.00 0.22
     0.44 0.85 0.22 1.00]
eigvals(ρ)

r = nearestPSDcor(ρ)
eigvals(r)